pyproject.toml
[project]
name = "project-mqtt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "paho-mqtt>=2.1.0",
    "pillow>=12.1.0",
    "pygame>=2.6.1",
]
